back/requirements.txt
Django>=3.1.0,<3.2.0
djangorestframework>=3.12.1,<3.13.0
psycopg2>=2.7.5,<2.8.0
pyotp>=2.4.0,<2.5.0
